+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.2.0
+- remove hash tree from storage (x)
+- add hash storage function (x)
+- add state history (x)
+- add state history auto trigger (x)
+---
 0.1.0
 - basic engine
 - basic action